[REF] sale_order_type: Split off stock dependent part
