+++ sale_order_type/data/default_type.xml
@@ -1,7 +1,7 @@
 <odoo noupdate="1">
     <record id="normal_sale_type" model="sale.order.type">
         <field name="name">Normal Order</field>
-        <field name="company_id" eval="False" />
+        <field name="company_id" eval="ref('base.main_company')" />
         <field name="sequence_id" ref="sale.seq_sale_order" />
     </record>
 </odoo>
