+++ sale_order_type/models/sale_order_type.py
@@ -13,17 +13,6 @@ class SaleOrderTypology(models.Model):
         seq_type = self.env.ref("sale.seq_sale_order")
         return [("code", "=", seq_type.code)]
 
-    @api.model
-    def _get_selection_picking_policy(self):
-        return self.env["sale.order"].fields_get(allfields=["picking_policy"])[
-            "picking_policy"
-        ]["selection"]
-
-    @api.model
-    def default_picking_policy(self):
-        default_dict = self.env["sale.order"].default_get(["picking_policy"])
-        return default_dict.get("picking_policy")
-
     name = fields.Char(required=True, translate=True)
     description = fields.Text(translate=True)
     sequence_id = fields.Many2one(
@@ -42,32 +31,24 @@ class SaleOrderTypology(models.Model):
     warehouse_id = fields.Many2one(
         comodel_name="stock.warehouse", string="Warehouse", check_company=True
     )
-    picking_policy = fields.Selection(
-        selection="_get_selection_picking_policy",
-        string="Shipping Policy",
-        default=default_picking_policy,
-    )
     company_id = fields.Many2one(
         comodel_name="res.company", default=lambda self: self.env.company, store=True,
     )
     payment_term_id = fields.Many2one(
-        comodel_name="account.payment.term", string="Payment Term", check_company=True
+        comodel_name="account.payment.term",
+        string="Payment Term",
+        sale_order_field="payment_term_id",
     )
     pricelist_id = fields.Many2one(
-        comodel_name="product.pricelist", strint="Pricelist", check_company=True
-    )
-    incoterm_id = fields.Many2one(comodel_name="account.incoterms", string="Incoterm")
-    route_id = fields.Many2one(
-        "stock.location.route",
-        string="Route",
-        domain=[("sale_selectable", "=", True)],
-        ondelete="restrict",
-        check_company=True,
+        comodel_name="product.pricelist",
+        strint="Pricelist",
+        sale_order_field="pricelist_id",
     )
     analytic_account_id = fields.Many2one(
         comodel_name="account.analytic.account",
         string="Analytic account",
         check_company=True,
+        sale_order_field="analytic_account_id",
     )
     active = fields.Boolean(default=True)
     quotation_validity_days = fields.Integer(string="Quotation Validity (Days)")
