+++ sale_order_type/readme/DESCRIPTION.rst
@@ -1,9 +1,7 @@
 This module adds a typology for the sales orders. In each different type, you
-can define, invoicing and refunding journal, a warehouse, a stock route,
-a sequence, the shipping policy, the invoicing policy, a payment term,
-a pricelist and an incoterm.
+can define a journal, a sequence, a payment term and a pricelist.
 
 You can see sale types as lines of business.
 
-You are able to select a sales order type by partner so that when you add a
+You are able to select a default sales order type by partner so that when you add a
 partner to a sales order it will get the related info to it.
