+++ sale_order_type/views/sale_order_type_view.xml
@@ -13,7 +13,7 @@
                         bg_color="bg-danger"
                         attrs="{'invisible': [('active', '=', True)]}"
                     />
-                    <group>
+                    <group name="header">
                         <div class="oe_title">
                             <label for="name" class="oe_edit_only" />
                             <h1>
@@ -21,21 +21,11 @@
                             </h1>
                         </div>
                     </group>
-                    <group>
-                        <group>
-                            <field
-                                name="warehouse_id"
-                                options="{'no_create': True}"
-                                groups="stock.group_stock_multi_warehouses"
-                            />
+                    <group name="main">
+                        <group name="left">
                             <field
                                 name="company_id"
                                 groups="base.group_multi_company"
-                            />
-                            <field
-                                name="route_id"
-                                attrs="{'invisible': [('warehouse_id', '=', False)]}"
-                                groups="stock.group_adv_location"
                             />
                             <field name="sequence_id" />
                             <field
@@ -43,22 +33,17 @@
                                 groups="analytic.group_analytic_accounting"
                             />
                         </group>
-                        <group>
+                        <group name="right">
                             <field name="journal_id" />
-                            <field name="picking_policy" />
                             <field name="quotation_validity_days" />
                             <field name="payment_term_id" />
                             <field
                                 name="pricelist_id"
                                 groups="product.group_product_pricelist"
                             />
-                            <field
-                                name="incoterm_id"
-                                groups="sale_stock.group_display_incoterm"
-                            />
                         </group>
                     </group>
-                    <group>
+                    <group name="footer">
                         <field name="description" nolabel="1" />
                     </group>
                 </sheet>
@@ -71,17 +56,10 @@
         <field name="arch" type="xml">
             <tree>
                 <field name="name" />
-                <field
-                    name="warehouse_id"
-                    groups="stock.group_stock_multi_warehouses"
-                />
-                <field name="route_id" groups="stock.group_adv_location" />
                 <field name="sequence_id" />
                 <field name="journal_id" />
-                <field name="picking_policy" />
                 <field name="payment_term_id" />
                 <field name="pricelist_id" groups="product.group_product_pricelist" />
-                <field name="incoterm_id" groups="sale_stock.group_display_incoterm" />
                 <field name="company_id" groups="base.group_multi_company" />
                 <field name="description" />
             </tree>
@@ -93,26 +71,12 @@
         <field name="arch" type="xml">
             <search string="Type">
                 <field name="name" />
-                <field
-                    name="warehouse_id"
-                    groups="stock.group_stock_multi_warehouses"
-                />
-                <field name="route_id" groups="stock.group_adv_location" />
                 <field name="sequence_id" />
                 <field name="journal_id" />
-                <field name="picking_policy" />
                 <field name="payment_term_id" />
                 <field name="pricelist_id" groups="product.group_product_pricelist" />
-                <field name="incoterm_id" groups="sale_stock.group_display_incoterm" />
                 <field name="company_id" groups="base.group_multi_company" />
                 <group expand="0" string="Group By">
-                    <filter
-                        name="warehouse"
-                        string="Warehouse"
-                        domain="[]"
-                        context="{'group_by':'warehouse_id'}"
-                        groups="stock.group_stock_multi_warehouses"
-                    />
                     <filter
                         name="entry_sequence"
                         string="Entry Sequence"
@@ -126,12 +90,6 @@
                         context="{'group_by':'journal_id'}"
                     />
                     <filter
-                        name="shipping_policy"
-                        string="Shipping Policy"
-                        domain="[]"
-                        context="{'group_by':'picking_policy'}"
-                    />
-                    <filter
                         name="payment_term"
                         string="Payment Term"
                         domain="[]"
@@ -143,13 +101,6 @@
                         domain="[]"
                         context="{'group_by':'pricelist_id'}"
                         groups="product.group_product_pricelist"
-                    />
-                    <filter
-                        name="incoterm"
-                        string="Incoterm"
-                        domain="[]"
-                        context="{'group_by':'incoterm_id'}"
-                        groups="sale_stock.group_display_incoterm"
                     />
                 </group>
             </search>
