+++ sale_order_type_stock/__manifest__.py
@@ -4,13 +4,11 @@
 # Copyright 2016 Lorenzo Battistini
 # Copyright 2016 Carlos Dauden <carlos.dauden@tecnativa.com>
 # Copyright 2018 David Vidal <david.vidal@tecnativa.com>
-# Copyright 2021 Therp BV <https://therp.nl>.
-# Copyright 2021 Sunflower IT < https://sunflowerweb.nl>.
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 {
-    "name": "Sale Order Type",
-    "version": "13.0.2.0.0",
+    "name": "Sale Order Type (with stock)",
+    "version": "13.0.1.0.0",
     "category": "Sales Management",
     "author": "Grupo Vermon,"
     "AvanzOSC,"
@@ -20,16 +18,8 @@
     "Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/sale-workflow",
     "license": "AGPL-3",
-    "depends": ["account", "sale_management"],
-    "demo": ["demo/sale_order_demo.xml"],
-    "data": [
-        "security/ir.model.access.csv",
-        "security/security.xml",
-        "views/sale_order_view.xml",
-        "views/sale_order_type_view.xml",
-        "views/account_move_views.xml",
-        "views/res_partner_view.xml",
-        "data/default_type.xml",
-    ],
+    "depends": ["sale_stock", "sale_order_type"],
+    "data": ["views/sale_order_type_view.xml", "data/default_type.xml"],
     "installable": True,
+    "auto_install": True,
 }
