+++ sale_order_type_stock/data/default_type.xml
@@ -1,0 +1,5 @@
+<odoo noupdate="1">
+    <record id="sale_order_type.normal_sale_type" model="sale.order.type">
+        <field name="warehouse_id" ref="stock.warehouse0" />
+    </record>
+</odoo>
