+++ sale_order_type_stock/models/__init__.py
@@ -1,0 +1,4 @@
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from . import sale_order_type
+from . import sale
