+++ sale_order_type_stock/readme/CONTRIBUTORS.rst
@@ -1,0 +1,42 @@
+* `Vermon <http://www.grupovermon.com>`_
+
+  * Carlos Sánchez Cifuentes <csanchez@grupovermon.com>
+
+* `AvanzOsc <http://avanzosc.es>`_
+
+  * Oihane Crucelaegui <oihanecrucelaegi@avanzosc.es>
+  * Ana Juaristi <anajuaristi@avanzosc.es>
+  * Daniel Campos <danielcampos@avanzosc.es>
+  * Ainara Galdona <ainaragaldona@avanzosc.es>
+
+* `Agile Business Group <https://www.agilebg.com>`_
+
+  * Lorenzo Battistini <lorenzo.battistini@agilebg.com>
+
+* `Niboo <https://www.niboo.be/>`_
+
+  * Samuel Lefever <sam@niboo.be>
+  * Pierre Faniel <pierre@niboo.be>
+
+* `Tecnativa <https://www.tecnativa.com>`_
+
+  * Pedro M. Baeza
+  * David Vidal
+  * Carlos Dauden
+
+* `Pesol <https://www.pesol.es>`_
+
+  * Angel Moya Pardo <angel.moya@pesol.es>
+  * Antonio J Rubio Lorente <antonio.rubio@pesol.es>
+
+* Rattapong Chokmasermkul <rattapongc@ecosoft.co.th>
+
+* `Druidoo <https://www.druidoo.io>`_
+
+  * Iván Todorovich <ivan.todorovich@druidoo.io>
+
+* `GSLab.it <https://www.gslab.it>`_
+
+  * Giovanni Serra <giovanni@gslab.it>
+
+Do not contact contributors directly about support or help with technical issues.
