+++ sale_order_type_stock/readme/DESCRIPTION.rst
@@ -1,0 +1,2 @@
+This module adds a warehouse, a stock route, a shipping policy and an incoterm
+to sale order types.
