+++ sale_order_type_stock/tests/__init__.py
@@ -1,0 +1,3 @@
+# License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
+
+from . import test_sale_order_type
