+++ setup/sale_order_type_stock/odoo/addons/sale_order_type_stock
@@ -1,0 +1,1 @@
+../../../../sale_order_type_stock
