+++ setup/sale_order_type_stock/setup.py
@@ -3,7 +3,4 @@ import setuptools
 setuptools.setup(
     setup_requires=['setuptools-odoo'],
     odoo_addon=True,
-    install_requires=[
-        "odoo13-addon-sale_order_type_stock",
-    ],
 )
